backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "ipython>=9.2.0",
    "markdown>=3.8",
    "psycopg[binary]>=3.2.9",
]

[dependency-groups]
dev = [
    "ruff>=0.11.11",
]
